refactor: reorganize bot into modules
